+++ sdk/storage/Azure.Storage.Files.Shares/src/ShareClient.cs
@@ -293,7 +293,9 @@ namespace Azure.Storage.Files.Shares
             ShareClientOptions options = default)
             : this(
                   shareUri: shareUri,
-                  authentication: credential.AsPolicy(options),
+                  authentication: credential.AsPolicy(
+                    options?.Audience != null ? options.Audience.Value.ToString() : ShareAudience.PublicAudience.ToString(),
+                    options),
                   options: options ?? new ShareClientOptions(),
                   storageSharedKeyCredential: default,
                   sasCredential: default,