+++ sdk/storage/Azure.Storage.Files.Shares/src/ShareDirectoryClient.cs
@@ -325,7 +325,9 @@ namespace Azure.Storage.Files.Shares
             ShareClientOptions options = default)
             : this(
                   directoryUri: directoryUri,
-                  authentication: credential.AsPolicy(options),
+                  authentication: credential.AsPolicy(
+                    options?.Audience != null ? options.Audience.Value.ToString() : ShareAudience.PublicAudience.ToString(),
+                    options),
                   options: options ?? new ShareClientOptions(),
                   storageSharedKeyCredential: null,
                   sasCredential: null,