+++ sdk/storage/Azure.Storage.Files.Shares/src/ShareServiceClient.cs
@@ -273,7 +273,9 @@ namespace Azure.Storage.Files.Shares
             ShareClientOptions options = default)
             : this(
                   serviceUri: serviceUri,
-                  authentication: credential.AsPolicy(options),
+                  authentication: credential.AsPolicy(
+                    options?.Audience != null ? options.Audience.Value.ToString() : ShareAudience.PublicAudience.ToString(),
+                    options),
                   options: options ?? new ShareClientOptions(),
                   sharedKeyCredential: null,
                   sasCredential: null,