+++ sdk/storage/Azure.Storage.Files.Shares/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 12.16.0-beta.2 (Unreleased)
 
 ### Features Added
+- Added support for ShareClientOptions.Audience
 
 ### Breaking Changes